requirements.txt
bcrypt==4.0.1
blinker==1.6.2
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.0.2
Flask-Admin==1.5.8
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-User==1.0.2.0
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.0.3
MarkupSafe==2.1.2
passlib==1.7.4
ply==3.11
pyasn1==0.5.0
pycparser==2.21
pycryptodomex==3.17
pysmi==0.3.4
pysnmp==4.4.12
SQLAlchemy==1.4.23
typing_extensions==4.5.0
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.15.0
email_validator
